backend/requirements.txt
django==3.1.0
djangorestframework==3.11.1
django-cors-headers==3.4.0
djangorestframework-simplejwt==4.4.0
django-filter==2.3.0
channels==2.4.0
channels-redis==3.0.1
redis==3.5.3